# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UFW Manger API",
    "description": "This is a set of APIs you can use for managing the UFW service.",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api"
    }
  ],
...
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Create new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth_body"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"username\":\"test\",\n    \"password\":\"Test1!\"\n}"
                },
                "1": {
                  "value": "{\n    \"username\":\"test\",\n    \"password\":\"Test123!\"\n}"
                },
                "2": {
                  "value": "{\n    \"username\":\"test\",\n    \"password\":\"Test123!\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Signed up successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\n  \"message\": \"your username registered, contact your system administrator to get your access\"\n}\n"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"The password needs to be more than 8 char and must include upper case, lower case, digits and symbols like Test123!\"}\n"
                  }
                }
              }
            }
          },
          "409": {
            "description": "The user already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"This username isn't available.\"}\n"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "You need to sign in to get api token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth_body"
              },
              "examples": {
                "0": {
                  "value": "{\n    \n}"
                },
                "1": {
                  "value": "{\n    \"username\":\"test\",\n    \"password\":\"Test12!\"\n}"
                },
                "2": {
                  "value": "{\n    \"username\":\"test\",\n    \"password\":\"Test123!\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "api token and sign in message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signin_response_200"
                },
                "examples": {
                  "0": {
                    "value": "{\"access_token\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJmcmVzaCI6ZmFsc2UsImlhdCI6MTYxOTM0NDcyMywianRpIjoiOWZlOWNkZDktNjE2MS00NDRjLTllNzEtYTExYjMxYzhmODY4IiwibmJmIjoxNjE5MzQ0NzIzLCJ0eXBlIjoiYWNjZXNzIiwic3ViIjoidGVzdCIsImV4cCI6MTc0Mzc2MDcyM30.7OjacjsL5RsfcIx46tmk-Xymc9n75ny8co8Z-ZMPf7M\",\"message\":\"Login succeeded.\"}\n"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Wrong username or password!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"Wrong username or password!\"}\n"
                  },
                  "1": {
                    "value": "{\"message\":\"Wrong username or password!\"}\n"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/change_password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Change password of the signed-in user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/change_password_body"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"current_password\":\"Test1234!\",\n    \"new_password\":\"Test123!\",\n    \"confirm_new_password\":\"Test123!\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user's password changed to the new password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"Your password changed successfully.\"}\n"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"Unauthorized user\"}\n"
                  },
                  "1": {
                    "value": "{\"message\":\"Unauthorized user\"}\n"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "admin",
              "manager",
              "user"
            ]
          }
        ]
      }
    },
    "/grant_authorization": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Admin can grant authorization for users",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/grant_authorization_body"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"username\":\"test\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authorization granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/grant_authorization_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":[\"test authorized with manager privileges\",\"date: 2021-04-25 14:59:49\"]}\n"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Authorization granted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/grant_authorization_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":[\"user 'test1' does not exist\",\"date: 2021-04-25 14:59:49\"]}\n"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/user_routes_response"
                },
                "examples": {
                  "0": {
                    "value": "{\"message\":\"Unauthorized user\"}\n"
                  },
                  "1": {
                    "value": "{\"message\":\"Unauthorized user\"}\n"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "admin"
            ]
          }
        ]
      }
    },
    "/get_user_privileges": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "To get user privileges",
        "responses": {
          "200": {
            "description": "Response with user privileges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_user_privileges_200"
                },
                "examples": {
                  "0": {
                    "value": "{\"result\": { \"isAdmin\": true,\"isManager\": true}, \"date\": \"2021-05-13 22:16:14\"}"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/unauthorized_user_403"
                },
                "examples": {
                  "0": {
                    "value": "{\"result\": \"Unauthorized user\", \"date\": \"2021-05-13 22:16:14\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "admin",
              "manager",
              "user"
            ]
          }
        ]
      }
    },
    "/get_users": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Admin can get users by id or part of the username",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID that needs to considered for filter",
            "required": false,
            "type": "integer"
          },
          {
            "name": "username",
            "in": "query",
            "description": "Username that needs to considered for filter",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "Number of records per page",
            "required": false,
            "type": "integer",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "List of users with ID or username as the one you send",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_users_list_response_200"
                },
                "examples": {
                  "0": {
                    "value": "{ \"result\": [{\"admin_privileges\": false,\"id\": 2,\"manager_privileges\": true,\"username\": \"test\"},{\"admin_privileges\": false,\"id\": 3, \"manager_privileges\": false,\"username\": \"test1\"}],\"total\": 2,\"date\": \"2021-05-13 22:32:18\" }"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No user found with this ID or username",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_users_list_response_200"
                },
                "examples": {
                  "0": {
                    "value": "{ \"result\": [],\"total\": 0,\"date\": \"2021-05-13 22:32:18\" }"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/list_unauthorized_user_403"
                },
                "examples": {
                  "0": {
                    "value": "{\"result\":\"Unauthorized user\",\"total\": 0, \"date\": \"2021-04-25 14:59:49\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "admin"
            ]
          }
        ]
      }
    },
    "/get_all_user": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Admin can get all users",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "perPage",
            "in": "query",
            "description": "Number of records per page",
            "required": false,
            "type": "integer",
            "default": 5
          }
        ],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/get_users_list_response_200"
                },
                "examples": {
                  "0": {
                    "value": "{ \"result\": [{\"admin_privileges\": true,\"id\": 1,\"manager_privileges\": true,\"username\": \"admin\"},{\"admin_privileges\": false,\"id\": 2, \"manager_privileges\": true,\"username\": \"test\"}],\"total\": 2,\"date\": \"2021-05-13 22:32:18\" }"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/list_unauthorized_user_403"
                },
                "examples": {
                  "0": {
                    "value": "{\"result\":\"Unauthorized user\",\"total\": 0, \"date\": \"2021-04-25 14:59:49\"}"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": [
              "admin"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "admin_privileges": {
            "type": "boolean"
          },
          "manager_privileges": {
            "type": "boolean"
          }
        }
      },
...
      "auth_body": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "grant_authorization_body": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "change_password_body": {
        "type": "object",
        "properties": {
          "confirm_new_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          },
          "current_password": {
            "type": "string"
          }
        }
      },
      "signin_response_200": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "user_routes_response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "grant_authorization_response": {
        "type": "object",
        "properties": {
          "message": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "get_users_list_response_200": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "total": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "get_user_privileges_200": {
        "type": "object",
        "properties": {
          "result": {
            "type": "object",
            "properties": {
              "isAdmin": {
                "type": "boolean"
              },
              "isManager": {
                "type": "boolean"
              }
            }
          },
          "date": {
            "type": "string"
          }
        }
      },
      "unauthorized_user_403": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "list_unauthorized_user_403": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}